Adding hero to the homepage
The hero consists of a header and a main body.
The header contains all the links to other sites including a logo, a
link as well.
Whereas the main body contains the main description of the website
including an image placeholder.
Nearly everything is done with flexboxes, but i am not quite satisfied
with the outcome, it can be designed much more cleaner.
The header and main text share the same distance (padding) to the border
of the homepage (by giving padding a % instead of a fixed px).